+++ Runtime/UI/LoadingSlider.cs
@@ -12,8 +12,8 @@ namespace ActionCode.SceneManagement
     {
         [SerializeField, Tooltip("The local Slider component.")]
         private Slider slider;
-        [SerializeField, Tooltip("The Scene Manager settings.")]
-        private SceneManagerSettings settings;
+        [SerializeField, Tooltip("The Scene Manager.")]
+        private SceneManager sceneManager;
 
         private void Reset()
         {
@@ -25,8 +25,8 @@ namespace ActionCode.SceneManagement
             slider.value = 50F;
         }
 
-        private void OnEnable() => settings.Transition.OnProgressChanged += HandleProgressChanged;
-        private void OnDisable() => settings.Transition.OnProgressChanged -= HandleProgressChanged;
+        private void OnEnable() => sceneManager.Transition.OnProgressChanged += HandleProgressChanged;
+        private void OnDisable() => sceneManager.Transition.OnProgressChanged -= HandleProgressChanged;
 
         private void HandleProgressChanged(float progress) => slider.value = progress;
     }