+++ Runtime/UI/LoadingText.cs
@@ -12,14 +12,14 @@ namespace ActionCode.SceneManagement
     {
         [SerializeField, Tooltip("The local Text component.")]
         private Text text;
-        [SerializeField, Tooltip("The Scene Manager settings.")]
-        private SceneManagerSettings settings;
+        [SerializeField, Tooltip("The Scene Manager.")]
+        private SceneManager sceneManager;
 
         public const string TEXT_FORMAT = "{0} %";
 
         private void Reset() => text = GetComponent<Text>();
-        private void OnEnable() => settings.Transition.OnProgressChanged += HandleProgressChanged;
-        private void OnDisable() => settings.Transition.OnProgressChanged -= HandleProgressChanged;
+        private void OnEnable() => sceneManager.Transition.OnProgressChanged += HandleProgressChanged;
+        private void OnDisable() => sceneManager.Transition.OnProgressChanged -= HandleProgressChanged;
 
         private void HandleProgressChanged(float progress) =>
             text.text = string.Format(TEXT_FORMAT, progress);