+++ CHANGELOG.md
@@ -11,12 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - IScreenFader interface
 - AnimationScreenFader component and prefab
 - ScreenFaderPool
-- SceneTransition
-- SceneTransitionData
+- SceneTransition and SceneTransitionData
 
 ### Changed
-- Scriptable Settings Provider for Scene Manager Setting 
+- Scriptable Settings Provider for Scene Manager 
 - Replace ScreenFaderCanvas -> CanvasScreenFader
+- SceneLoadingSettings -> SceneManager
 
 ### Removed
 - Animation Parameters package